% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_leverage.R
\name{algo_leverage}
\alias{algo_leverage}
\title{Leveraging techniques with two different methods (uniform and hat).
Returns the coefficients from a linear model call.}
\usage{
algo_leverage(y, X, size = 10, method = "hat")
}
\arguments{
\item{y}{response (vector or matrix)}

\item{X}{predictors (matrix)}

\item{size}{size option for uniform sub-sampling}

\item{method}{either uniform (random subsampling) or hat (SVD-based) leveraging.
See Algorithm 1 of Ma and Sun, "Leveraging for big data regression," WIREs Comp. Stat. 2014.}
}
\description{
Leveraging techniques with two different methods (uniform and hat).
Returns the coefficients from a linear model call.
}
\examples{
X = matrix(rnorm(100), ncol = 2)
y = rnorm(50)
algo_leverage(y, X, method = "hat")

}
