% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet_coord.R
\name{elnet_coord}
\alias{elnet_coord}
\title{implementation of elastic net coordinate descent as described in Statistical Learning
with Sparsity, Hastie, Tibshirani, and Wainwright.
eg. hat{beta} = argmin_{beta} ||y - x beta||_2^2 + lambda(alpha ||beta||_1 + (1-alpha) ||beta||_2)}
\usage{
elnet_coord(y, x, lambda = 1, alpha = 0, maxSteps = 100,
  minTol = 1e-08)
}
\arguments{
\item{y}{response vector}

\item{x}{input feature matrix}

\item{lambda}{shrinkage parameter}

\item{alpha}{sparsity parameter for elastic net (default: 0, so lasso)}

\item{maxSteps}{maximum steps for iterative coordinate descent (default: 100)}

\item{minTol}{minimum convergence tolerance (default: 1e-8)}
}
\description{
returns the parameters estimated
}
\examples{
data = gen_fs_system()
elnet_coord(data$y, data$x, lambda = 0.5, alpha = 0.5)
}
