% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet_coord.R
\name{gen_fs_system}
\alias{gen_fs_system}
\title{generates the simulated data for a sparse system}
\usage{
gen_fs_system(n = 20, p = 20, noise = 1, seed = 2019)
}
\arguments{
\item{n}{number of data points}

\item{p}{dimensionality of system}

\item{noise}{noise in system}

\item{seed}{enables setting seed for reproducibility}
}
\description{
generates the simulated data for a sparse system
}
\examples{
output = gen_fs_system()
output$y

}
