% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ols.R
\name{gen_tridiag_mat}
\alias{gen_tridiag_mat}
\title{generates a tri-diagonal matrix of size n}
\usage{
gen_tridiag_mat(alpha = 1, n = 100)
}
\arguments{
\item{alpha}{diagonal of matrix}

\item{n}{size of matrix}
}
\description{
generates a tri-diagonal matrix of size n
}
\examples{
A = gen_tridiag_mat(4, 10)
A

}
